Code/hardware/requirements.txt
Adafruit-Blinka==7.2.1
adafruit-circuitpython-dht==3.7.2
Adafruit-PlatformDetect==3.22.1
Adafruit-PureIO==1.1.9
adafruit-python-shell==1.3.3
args==0.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
clint==0.5.1
idna==3.3
psutil==5.9.0
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
requests==2.27.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
smbus==1.1.post2
sysv-ipc==1.1.0
urllib3==1.26.9
